reloc: one module per prior; eval replaces verbose_eval_logging
